some more refectoring
