+++ src/app/components/core/asset-readings/assets/assets.component.html
@@ -120,7 +120,7 @@
       </table>
     </div>
     <app-asset-summary></app-asset-summary>
-    <app-chart-modal></app-chart-modal>
+    <app-readings-graph></app-readings-graph>
     <div *ngIf='totalPagesCount > 1' class="pagination-section">
       <app-pagination [count]='recordCount' [page]="page" [perPage]='limit' [totalPage]='totalPagesCount' (goPrev)="onPrev()" (goNext)="onNext()"
         (goFirst)="onFirst()" (goLast)="onLast()" (goPage)="goToPage($event)">
