+++ src/app/components/core/asset-readings/readings-graph/readings-graph.component.css
