+++ src/app/components/core/asset-readings/readings-graph/readings-graph.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ReadingsGraphComponent } from './readings-graph.component';
+
+describe('ReadingsGraphComponent', () => {
+  let component: ReadingsGraphComponent;
+  let fixture: ComponentFixture<ReadingsGraphComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ ReadingsGraphComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ReadingsGraphComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
