+++ src/app/components/core/scheduler/list-schedules/list-schedules.component.css
@@ -1,0 +1,15 @@
+table.scroll {
+  height: 300px;
+}
+
+.fa-minus-circle {
+  color: rgb(178, 34, 34)
+}
+
+.fade {
+  background-color:#f5f5f5
+}
+
+.fa::before {
+  display: inline-block;
+}
