+++ src/app/components/core/scheduler/list-schedules/list-schedules.component.html
@@ -1,0 +1,73 @@
+<div class="card">
+  <header class="card-header">
+    <p class="card-header-title">
+      Schedules &nbsp;
+      <button (click)="getSchedules()" class="button is-small">
+        <i class="fa fa-refresh" aria-hidden="true"></i>
+      </button>
+    </p>
+    <div>
+      <a class="button is-light" (click)="openCreateSchedulerModal()">Create &nbsp;
+        <i class="fa fa-plus" aria-hidden="true"></i>
+      </a>
+    </div>
+  </header>
+  <div class="card-content">
+    <table class="table scroll is-responsive">
+      <thead id="head">
+        <tr>
+          <th>Name</th>
+          <th>Type</th>
+          <th>Day
+            <!-- <span class="tooltip is-tooltip-right" data-tooltip="ISO Days (1-7 Mon-Sun)">
+                    <i class="fa fa-info-circle"></i>
+                  </span> -->
+          </th>
+          <th>Time</th>
+          <th>Interval</th>
+          <th>Exclusive</th>
+          <th></th>
+        </tr>
+      </thead>
+      <tr *ngFor="let schedule of scheduleData" [ngClass]="{'fade': !schedule.enabled}" id={{schedule.name}}>
+        <td>
+          <small> {{schedule.name}} </small>
+        </td>
+        <td>
+          <small> {{schedule.type}}</small>
+        </td>
+        <td>
+          <small> {{getISODay(schedule.day)}} </small>
+        </td>
+        <td>
+          <small> {{schedule.time}}</small>
+        </td>
+        <td>
+          <small> {{schedule.repeat}}</small>
+        </td>
+        <td>
+          <small> {{schedule.exclusive}}</small>
+        </td>
+        <td>
+          <a (click)="editSchedule(schedule.id)" class="button is-text" name={{schedule.name}}>
+            <i class="fa fa-pencil-square-o" aria-hidden="true"></i>
+          </a>
+        </td>
+        <td>
+          <a name={{schedule.name}} (click)="openModal(schedule.id, schedule.name, 'Are you sure, You want to delete the schedule', 'delete')"
+            class="button is-text">
+            <i class="fa fa-minus-circle" aria-hidden="true"></i>
+          </a>
+        </td>
+        <td>
+          <button name={{schedule.name}} *ngIf='schedule.enabled' class="button is-text" (click)="openModal(schedule.id, schedule.name, 'Are you sure, You want to disable the schedule', 'disable')">disable</button>
+          <button name={{schedule.name}} *ngIf='!schedule.enabled' class="button is-text" (click)="openModal(schedule.id, schedule.name, 'Are you sure, You want to enable the schedule', 'enable')">enable</button>
+        </td>
+      </tr>
+    </table>
+  </div>
+</div>
+<app-alert-dialog (enable)='enableSchedule($event)' (disable)='disableSchedule($event)' (delete)='deleteSchedule($event)'
+  [childData]='childData'></app-alert-dialog>
+<app-update-schedule (notify)='onNotify()' [childData]='updateScheduleData'></app-update-schedule>
+<app-create-schedule (notify)='onNotify()' (process)='setScheduleProcess($event)' (type)='setScheduleType($event)'></app-create-schedule>
